requirements.txt
matplotlib>=3.0.0
pandas>=0.23.4
msgpack-numpy>=0.4.3.2
spacy>=2.0.12
scipy>=1.1.0
scikit-learn>=0.20.0
nltk>=3.4
dill>=0.2.9
torch>=0.12
clean-text>=0.1.1
joblib>=0.13.2